* transcode.c (transcode_loop): suppressed a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@27386 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ transcode.c
@@ -2280,7 +2280,7 @@ transcode_loop(const unsigned char **in_pos, unsigned char **out_pos,
 	    StringValue(rep);
 	    ret = rb_econv_insert_output(ec, (const unsigned char *)RSTRING_PTR(rep),
 		    RSTRING_LEN(rep), rb_enc_name(rb_enc_get(rep)));
-	    if (ret == -1) {
+	    if ((int)ret == -1) {
 		rb_raise(rb_eArgError, "too big fallback string");
 	    }
 	    goto resume;
